# backend/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Agency Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /setter-activities:
    get:
      operationId: listSetterActivities
      tags: [setter-activities]
      summary: List setter activities
      parameters:
        - name: setterId
          in: query
          schema:
            type: ["integer", "null"]
        - name: dateFrom
          in: query
          schema:
            type: ["string", "null"]
        - name: dateTo
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of setter activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SetterActivity"
    post:
      operationId: createSetterActivity
      tags: [setter-activities]
      summary: Log setter activity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetterActivityInput"
      responses:
        "201":
          description: Setter activity created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetterActivity"
...
  /setter-activities/{id}:
    patch:
      operationId: updateSetterActivity
      tags: [setter-activities]
      summary: Update setter activity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetterActivityUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetterActivity"
    delete:
      operationId: deleteSetterActivity
      tags: [setter-activities]
      summary: Delete setter activity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Closer Activities
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # Auth
...
    SetterActivity:
      type: object
      required: [id, setterId, date, outreachCount, callsBooked, showRate, createdAt]
      properties:
        id:
          type: integer
        setterId:
          type: integer
        setterName:
          type: ["string", "null"]
        date:
          type: string
        outreachCount:
          type: integer
        callsBooked:
          type: integer
        showRate:
          type: number
        responseRate:
          type: ["number", "null"]
        platform:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    SetterActivityInput:
      type: object
      required: [setterId, date, outreachCount, callsBooked, showRate]
      properties:
        setterId:
          type: integer
        date:
          type: string
        outreachCount:
          type: integer
        callsBooked:
          type: integer
        showRate:
          type: number
        responseRate:
          type: number
        platform:
          type: string
        notes:
          type: string
    SetterActivityUpdate:
      type: object
      properties:
        outreachCount:
          type: integer
        callsBooked:
          type: integer
        showRate:
          type: number
        responseRate:
          type: number
        platform:
          type: string
        notes:
          type: string
...